+++ packages/espressocash_app/lib/features/country_picker/widgets/country_picker.dart
@@ -1,19 +1,21 @@
-import 'package:dfunc/dfunc.dart';
 import 'package:flutter/material.dart';
 
 import '../../../l10n/l10n.dart';
+import '../../../ui/colors.dart';
+import '../../../ui/picker_screen.dart';
 import '../models/country.dart';
-import '../screens/country_picker_screen.dart';
 
 class CountryPicker extends StatelessWidget {
   const CountryPicker({
     super.key,
     this.country,
+    this.placeholder,
     this.backgroundColor = Colors.black,
     required this.onSubmitted,
   });
 
   final Country? country;
+  final String? placeholder;
   final Color backgroundColor;
   final ValueSetter<Country> onSubmitted;
 
@@ -26,24 +28,28 @@ class CountryPicker extends StatelessWidget {
         child: ListTile(
           contentPadding: const EdgeInsets.symmetric(horizontal: 24),
           onTap: () async {
-            Country? country;
-            await CountryPickerScreen.push(
-              context,
+            await CustomPickerScreen.push<Country>(
+              context: context,
+              title: context.l10n.selectCountryTitle,
+              items: Country.all,
               initial: country,
-              onTap: (updated, _) async => country = updated,
+              itemBuilder: (context, country, {required bool selected}) => Text(
+                country.name,
+                style: TextStyle(
+                  fontSize: selected ? 19 : 17,
+                  color: Colors.white,
+                ),
+              ),
+              onTap: (country, context) async => onSubmitted(country),
             );
-
-            if (context.mounted) {
-              country?.let(onSubmitted);
-            }
           },
           title: Text(
-            country?.name ?? context.l10n.countryOfResidence,
-            style: const TextStyle(
+            country?.name ?? placeholder ?? context.l10n.countryOfResidence,
+            style: TextStyle(
               fontWeight: FontWeight.normal,
               fontSize: 16,
-              color: Colors.white,
               height: 1.2,
+              color: country != null ? Colors.white : CpColors.greyColor,
             ),
           ),
           trailing: const Icon(