+++ packages/espressocash_app/lib/features/feature_flags/data/feature_flags_manager.dart
@@ -4,13 +4,17 @@ import 'package:firebase_remote_config/firebase_remote_config.dart';
 import 'package:get_it/get_it.dart';
 import 'package:injectable/injectable.dart';
 
-@singleton
+import '../../accounts/auth_scope.dart';
+import '../../ambassador/data/ambassador_service.dart';
+
+@Singleton(scope: authScope)
 class FeatureFlagsManager implements Disposable {
-  FeatureFlagsManager() {
+  FeatureFlagsManager(this._ambassadorService) {
     _init();
   }
 
   final _remoteConfig = FirebaseRemoteConfig.instance;
+  final AmbassadorService _ambassadorService;
 
   StreamSubscription<void>? _subscription;
 
@@ -29,12 +33,11 @@ class FeatureFlagsManager implements Disposable {
   bool isMoneygramAccessEnabled() =>
       _remoteConfig.getBool(FeatureFlag.moneygram.name);
 
-  bool isBrijEnabled() => _remoteConfig.getBool(FeatureFlag.brij.name);
+  bool isBrijEnabled() =>
+      _remoteConfig.getBool(FeatureFlag.brij.name) ||
+      _ambassadorService.value.isAmbassador;
 
   bool isBrijDemoEnabled() => _remoteConfig.getBool(FeatureFlag.brijDemo.name);
-
-  bool isScalexBrijEnabled() =>
-      _remoteConfig.getBool(FeatureFlag.scalexBrij.name);
 
   @override
   void onDispose() {
@@ -42,4 +45,4 @@ class FeatureFlagsManager implements Disposable {
   }
 }
 
-enum FeatureFlag { moneygram, brij, brijDemo, scalexBrij }
+enum FeatureFlag { moneygram, brij, brijDemo }