+++ packages/espressocash_app/lib/features/kyc_sharing/screens/personal_information_screen.dart
@@ -7,43 +7,36 @@ import '../../../ui/bottom_button.dart';
 import '../../../ui/colors.dart';
 import '../../../ui/dob_text_field.dart';
 import '../../../ui/loader.dart';
-import '../../../ui/radio_button.dart';
 import '../../../ui/snackbar.dart';
 import '../../country_picker/models/country.dart';
 import '../../country_picker/widgets/country_picker.dart';
-import '../models/document_type.dart';
 import '../services/kyc_service.dart';
 import '../utils/kyc_utils.dart';
-import '../widgets/document_picker.dart';
-import '../widgets/kyc_header.dart';
 import '../widgets/kyc_page.dart';
 import '../widgets/kyc_text_field.dart';
 
-class BasicInformationScreen extends StatefulWidget {
-  const BasicInformationScreen({super.key});
+class PersonalInformationScreen extends StatefulWidget {
+  const PersonalInformationScreen({super.key});
 
   static Future<bool> push(BuildContext context) => Navigator.of(context)
       .push<bool>(
         MaterialPageRoute(
-          builder: (context) => const BasicInformationScreen(),
+          builder: (context) => const PersonalInformationScreen(),
         ),
       )
       .then((result) => result ?? false);
 
   @override
-  State<BasicInformationScreen> createState() => _BasicInformationScreenState();
+  State<PersonalInformationScreen> createState() =>
+      _PersonalInformationScreenState();
 }
 
-class _BasicInformationScreenState extends State<BasicInformationScreen> {
-  final _idNumberController = TextEditingController();
+class _PersonalInformationScreenState extends State<PersonalInformationScreen> {
   final _firstNameController = TextEditingController();
   final _lastNameController = TextEditingController();
   final _dobController = TextEditingController();
 
-  bool _isShareData = false;
-
-  Country? _country;
-  DocumentType? _idType;
+  Country? _citizenship;
 
   bool get _isValid {
     final DateTime? dob = _parseDate(_dobController.text);
@@ -52,10 +45,7 @@ class _BasicInformationScreenState extends State<BasicInformationScreen> {
         _lastNameController.text.trim().isNotEmpty &&
         dob != null &&
         !dob.isAfter(DateTime.now()) &&
-        _idNumberController.text.trim().isNotEmpty &&
-        _isShareData &&
-        _idType != null &&
-        _country != null;
+        _citizenship != null;
   }
 
   @override
@@ -66,7 +56,6 @@ class _BasicInformationScreenState extends State<BasicInformationScreen> {
 
   @override
   void dispose() {
-    _idNumberController.dispose();
     _firstNameController.dispose();
     _lastNameController.dispose();
     _dobController.dispose();
@@ -77,14 +66,11 @@ class _BasicInformationScreenState extends State<BasicInformationScreen> {
   void _initializeUserData() {
     final user = sl<KycSharingService>().value;
     final dob = user?.dob;
-    _country = Country.findByCode(user?.countryCode ?? '');
+    _citizenship = Country.findByCode(user?.citizenshipCode ?? '');
     _firstNameController.text = user?.firstName ?? '';
     _lastNameController.text = user?.lastName ?? '';
     _dobController.text =
         dob != null ? DateFormat('dd/MM/yyyy').format(dob) : '';
-    _idType = user?.documentType?.toDocumentType();
-    _idNumberController.text = user?.documentNumber ?? '';
-    _isShareData = user?.firstName?.isNotEmpty ?? false;
   }
 
   Future<void> _handleSubmitted() async {
@@ -93,20 +79,17 @@ class _BasicInformationScreenState extends State<BasicInformationScreen> {
       () async {
         try {
           final DateTime? dob = _parseDate(_dobController.text);
-          final countryCode = _country?.code;
-          final idTypeValue = _idType?.value;
+          final countryCode = _citizenship?.code;
 
-          if (countryCode == null || idTypeValue == null) {
+          if (countryCode == null) {
             throw Exception();
           }
 
-          await sl<KycSharingService>().updateBasicInfo(
+          await sl<KycSharingService>().updatePersonalInfo(
             firstName: _firstNameController.text,
             lastName: _lastNameController.text,
             dob: dob,
-            countryCode: countryCode,
-            idType: _idType,
-            idNumber: _idNumberController.text,
+            citizenshipCode: countryCode,
           );
 
           if (!mounted) return false;
@@ -141,21 +124,20 @@ class _BasicInformationScreenState extends State<BasicInformationScreen> {
 
   @override
   Widget build(BuildContext context) => KycPage(
+        title: 'Personal Information'.toUpperCase(),
         children: [
-          KycHeader(title: context.l10n.idVerificationTitle.toUpperCase()),
-          const SizedBox(height: 8),
           KycTextField(
             controller: _firstNameController,
             inputType: TextInputType.name,
             placeholder: context.l10n.firstName,
           ),
-          const _Divider(),
+          const SizedBox(height: 10),
           KycTextField(
             controller: _lastNameController,
             inputType: TextInputType.name,
             placeholder: context.l10n.lastName,
           ),
-          const _Divider(),
+          const SizedBox(height: 10),
           CpDobTextField(
             controller: _dobController,
             placeholder: context.l10n.dateOfBirth,
@@ -163,42 +145,9 @@ class _BasicInformationScreenState extends State<BasicInformationScreen> {
           const SizedBox(height: 10),
           CountryPicker(
             backgroundColor: CpColors.blackGreyColor,
-            country: _country,
-            onSubmitted: (country) => setState(() => _country = country),
-          ),
-          const SizedBox(height: 10),
-          DocumentPicker(
-            type: _idType,
-            onSubmitted: (idType) => setState(() => _idType = idType),
-          ),
-          const _Divider(),
-          KycTextField(
-            controller: _idNumberController,
-            inputType: TextInputType.text,
-            placeholder: context.l10n.idNumber,
-          ),
-          const _Divider(),
-          GestureDetector(
-            onTap: () => setState(() => _isShareData = !_isShareData),
-            child: Row(
-              children: [
-                CpRadioButton(
-                  value: _isShareData,
-                  onChanged: (value) => setState(() => _isShareData = value),
-                ),
-                const SizedBox(width: 12),
-                Expanded(
-                  child: Text(
-                    context.l10n.allowShareDataText,
-                    style: const TextStyle(
-                      fontSize: 14,
-                      height: 1.5,
-                      letterSpacing: 0.19,
-                    ),
-                  ),
-                ),
-              ],
-            ),
+            placeholder: 'Country of Citizenship',
+            country: _citizenship,
+            onSubmitted: (country) => setState(() => _citizenship = country),
           ),
           const SizedBox(height: 28),
           const Spacer(),
@@ -207,7 +156,6 @@ class _BasicInformationScreenState extends State<BasicInformationScreen> {
               _firstNameController,
               _lastNameController,
               _dobController,
-              _idNumberController,
             ]),
             builder: (context, child) => CpBottomButton(
               horizontalPadding: 16,
@@ -217,11 +165,4 @@ class _BasicInformationScreenState extends State<BasicInformationScreen> {
           ),
         ],
       );
-}
-
-class _Divider extends StatelessWidget {
-  const _Divider();
-
-  @override
-  Widget build(BuildContext context) => const SizedBox(height: 12);
 }