+++ packages/espressocash_app/lib/features/ramp/widgets/partner_config.dart
@@ -20,17 +20,11 @@ IList<RampPartner> getOnRampPartners(String? countryCode) {
     partners.add(RampPartner.kado);
   }
 
-  final isBrijEnabled = sl<FeatureFlagsManager>().isBrijEnabled();
+  final isBrijPartnerEnabled = sl<FeatureFlagsManager>().isBrijEnabled() &&
+      sl<FeatureFlagsManager>().isBrijDemoEnabled();
 
-  if (isBrijEnabled && _brijCountries.contains(countryCode)) {
+  if (isBrijPartnerEnabled && _brijCountries.contains(countryCode)) {
     partners.add(RampPartner.brij);
-  }
-
-  if (_scalexCountries.contains(countryCode)) {
-    final isScalexBrijEnabled = sl<FeatureFlagsManager>().isScalexBrijEnabled();
-
-    partners
-        .add(isScalexBrijEnabled ? RampPartner.scalexBrij : RampPartner.scalex);
   }
 
   partners.add(RampPartner.rampNetwork);
@@ -58,17 +52,11 @@ IList<RampPartner> getOffRampPartners(String? countryCode) {
     partners.add(RampPartner.coinflow);
   }
 
-  final isBrijEnabled = sl<FeatureFlagsManager>().isBrijEnabled();
+  final isBrijPartnerEnabled = sl<FeatureFlagsManager>().isBrijEnabled() &&
+      sl<FeatureFlagsManager>().isBrijDemoEnabled();
 
-  if (isBrijEnabled && _brijCountries.contains(countryCode)) {
+  if (isBrijPartnerEnabled && _brijCountries.contains(countryCode)) {
     partners.add(RampPartner.brij);
-  }
-
-  if (_scalexCountries.contains(countryCode)) {
-    final isScalexBrijEnabled = sl<FeatureFlagsManager>().isScalexBrijEnabled();
-
-    partners
-        .add(isScalexBrijEnabled ? RampPartner.scalexBrij : RampPartner.scalex);
   }
 
   return IList(partners);
@@ -87,8 +75,6 @@ const _coinflowCountries = {
   'DE', 'GR', 'HU', 'IS', 'IE', 'IT', 'LV', 'LI', 'LT', 'LU', 'MT', 'MC',
   'NL', 'NO', 'PL', 'PT', 'RO', 'SM', 'SK', 'SI', 'ES', 'SE', 'CH', 'US',
 };
-
-const _scalexCountries = {'NG'};
 
 const _moneygramOnRampCountries = {
   'AO', 'AR', 'AU', 'BH', 'BF', 'KH', 'CA', 'CV', 'CL', 'CD', 'CI', 'DK', //