+++ packages/espressocash_app/lib/ui/picker_screen.dart
@@ -1,123 +1,160 @@
 import 'package:dfunc/dfunc.dart';
 import 'package:flutter/material.dart';
 
-import '../../../l10n/l10n.dart';
-import '../../../ui/app_bar.dart';
-import '../../../ui/colors.dart';
-import '../../../ui/page_spacer_wrapper.dart';
-import '../../../ui/text_field.dart';
-import '../../../ui/theme.dart';
-import '../models/country.dart';
+import '../../l10n/l10n.dart';
+import '../../ui/app_bar.dart';
+import '../../ui/colors.dart';
+import '../../ui/text_field.dart';
+import '../../ui/theme.dart';
+import 'page_spacer_wrapper.dart';
 
-typedef CountryOnTap = Future<void> Function(
-  Country updatedCountry,
+typedef ItemBuilder<T> = Widget Function(
+  BuildContext context,
+  T item, {
+  required bool selected,
+});
+
+typedef ItemFilter<T> = bool Function(T item, String searchQuery);
+typedef ItemTapCallback<T> = Future<void> Function(
+  T selected,
   BuildContext context,
 );
 
-class CountryPickerScreen extends StatelessWidget {
-  const CountryPickerScreen({
+class CustomPickerScreen<T> extends StatelessWidget {
+  const CustomPickerScreen({
     super.key,
+    required this.items,
+    required this.itemBuilder,
+    required this.title,
     this.initial,
     this.onTap,
-    this.showDialCode = false,
+    this.filterItem,
+    this.searchPlaceholder,
   });
 
-  static Future<void> open(
+  static Future<void> open<T>(
     BuildContext context, {
-    Country? initial,
-    CountryOnTap? onTap,
     NavigatorState? navigator,
-    bool showDialCode = false,
+    required String title,
+    required List<T> items,
+    required ItemBuilder<T> itemBuilder,
+    T? initial,
+    Future<void> Function(T selected, BuildContext context)? onTap,
+    ItemFilter<T>? filterItem,
+    String? searchPlaceholder,
   }) =>
       (navigator ?? Navigator.of(context, rootNavigator: true))
-          .pushAndRemoveUntil<Country>(
+          .pushAndRemoveUntil<T>(
         PageRouteBuilder(
-          pageBuilder: (context, _, __) => CountryPickerScreen(
+          pageBuilder: (context, _, __) => CustomPickerScreen<T>(
+            title: title,
+            items: items,
+            itemBuilder: itemBuilder,
             initial: initial,
             onTap: onTap,
-            showDialCode: showDialCode,
+            filterItem: filterItem,
+            searchPlaceholder: searchPlaceholder,
           ),
           transitionDuration: Duration.zero,
         ),
         F,
       );
 
-  static Future<void> push(
-    BuildContext context, {
-    Country? initial,
-    CountryOnTap? onTap,
-    bool showDialCode = false,
+  static Future<void> push<T>({
+    required BuildContext context,
+    required String title,
+    required List<T> items,
+    required ItemBuilder<T> itemBuilder,
+    T? initial,
+    Future<void> Function(T selected, BuildContext context)? onTap,
+    ItemFilter<T>? filterItem,
+    String? searchPlaceholder,
   }) =>
       Navigator.of(context).push<void>(
         MaterialPageRoute(
-          builder: (context) => CountryPickerScreen(
+          builder: (context) => CustomPickerScreen<T>(
+            title: title,
+            items: items,
+            itemBuilder: itemBuilder,
             initial: initial,
             onTap: onTap,
-            showDialCode: showDialCode,
+            filterItem: filterItem,
+            searchPlaceholder: searchPlaceholder,
           ),
         ),
       );
 
-  final Country? initial;
-  final CountryOnTap? onTap;
-  final bool showDialCode;
+  final String title;
+  final List<T> items;
+  final ItemBuilder<T> itemBuilder;
+  final T? initial;
+  final ItemTapCallback<T>? onTap;
+  final ItemFilter<T>? filterItem;
+  final String? searchPlaceholder;
 
   @override
   Widget build(BuildContext context) => CpTheme.dark(
         child: Scaffold(
           backgroundColor: CpColors.deepGreyColor,
           appBar: CpAppBar(
-            title: Text(context.l10n.selectCountryTitle.toUpperCase()),
+            title: Text(title.toUpperCase()),
           ),
           body: _Wrapper(
-            child: _Content(
+            child: _Content<T>(
+              items: items,
+              itemBuilder: itemBuilder,
               initial: initial,
               onTap: onTap,
-              showDialCode: showDialCode,
+              filterItem: filterItem,
+              searchPlaceholder: searchPlaceholder,
             ),
           ),
         ),
       );
 }
 
-class _Content extends StatefulWidget {
+class _Content<T> extends StatefulWidget {
   const _Content({
+    required this.items,
+    required this.itemBuilder,
     this.initial,
-    required this.onTap,
-    required this.showDialCode,
+    this.onTap,
+    this.filterItem,
+    this.searchPlaceholder,
   });
 
-  final Country? initial;
-  final CountryOnTap? onTap;
-  final bool showDialCode;
+  final List<T> items;
+  final ItemBuilder<T> itemBuilder;
+  final T? initial;
+  final ItemTapCallback<T>? onTap;
+  final ItemFilter<T>? filterItem;
+  final String? searchPlaceholder;
 
   @override
-  State<_Content> createState() => _ContentState();
+  State<_Content<T>> createState() => _ContentState<T>();
 }
 
-class _ContentState extends State<_Content> {
+class _ContentState<T> extends State<_Content<T>> {
   final TextEditingController _searchController = TextEditingController();
   final ScrollController _scrollController = ScrollController();
 
-  Country? _selectedCountry;
+  T? _selectedItem;
   String _searchText = '';
-
-  final _countries = Country.all;
 
   @override
   void initState() {
     super.initState();
 
-    _selectedCountry = widget.initial;
+    _selectedItem = widget.initial;
 
     _searchController.addListener(() {
       setState(() => _searchText = _searchController.text);
     });
 
-    final country = _selectedCountry;
-    if (country != null) {
+    final item = _selectedItem;
+    if (item != null) {
       WidgetsBinding.instance.addPostFrameCallback((_) {
-        final index = _countries.indexOf(country);
+        final index = widget.items.indexOf(item);
         final centerOffset = ((context.size?.height ?? 0) - _tileHeight) / 2.5;
         final offset = index * _tileHeight - centerOffset;
         _scrollController.jumpTo(offset);
@@ -134,18 +171,12 @@ class _ContentState extends State<_Content> {
 
   @override
   Widget build(BuildContext context) {
-    final filteredCountries = _countries.where((country) {
-      final nameMatches =
-          country.name.toLowerCase().contains(_searchText.toLowerCase());
+    final filteredItems = widget.items.where((item) {
+      final filter = widget.filterItem;
 
-      final codeMatches =
-          country.code.toLowerCase().contains(_searchText.toLowerCase());
-
-      final dialCodeMatches = widget.showDialCode
-          ? country.dialCode.toLowerCase().contains(_searchText.toLowerCase())
-          : false;
-
-      return nameMatches || codeMatches || dialCodeMatches;
+      return filter != null
+          ? filter(item, _searchText)
+          : item.toString().toLowerCase().contains(_searchText.toLowerCase());
     }).toList();
 
     return Column(
@@ -158,13 +189,14 @@ class _ContentState extends State<_Content> {
             padding: const EdgeInsets.symmetric(vertical: 16, horizontal: 24),
             fontSize: 16,
             border: CpTextFieldBorder.stadium,
-            placeholder: context.l10n.searchPlaceholder,
+            placeholder:
+                widget.searchPlaceholder ?? context.l10n.searchPlaceholder,
             backgroundColor: CpColors.blackTextFieldBackgroundColor,
             textColor: Colors.white,
             inputType: TextInputType.text,
           ),
         ),
-        if (filteredCountries.isEmpty)
+        if (filteredItems.isEmpty)
           Expanded(
             child: Center(
               child: Text(
@@ -185,11 +217,11 @@ class _ContentState extends State<_Content> {
                 right: 20,
                 bottom: MediaQuery.paddingOf(context).bottom,
               ),
-              itemCount: filteredCountries.length,
+              itemCount: filteredItems.length,
               itemExtent: _tileHeight,
               itemBuilder: (BuildContext context, int index) {
-                final country = filteredCountries[index];
-                final selected = country == _selectedCountry;
+                final item = filteredItems[index];
+                final selected = item == _selectedItem;
 
                 return DecoratedBox(
                   decoration: selected
@@ -200,35 +232,12 @@ class _ContentState extends State<_Content> {
                       : const BoxDecoration(),
                   child: ListTile(
                     dense: true,
-                    title: Row(
-                      children: [
-                        if (widget.showDialCode)
-                          SizedBox(
-                            width: 70,
-                            child: Text(
-                              country.dialCode,
-                              style: TextStyle(
-                                fontSize: selected ? 19 : 17,
-                                color: CpColors.yellowColor,
-                              ),
-                            ),
-                          ),
-                        Expanded(
-                          child: Text(
-                            country.name,
-                            style: TextStyle(
-                              fontSize: selected ? 19 : 17,
-                              color: Colors.white,
-                            ),
-                          ),
-                        ),
-                      ],
-                    ),
+                    title:
+                        widget.itemBuilder(context, item, selected: selected),
                     selectedColor: Colors.white,
                     shape: selected ? const StadiumBorder() : null,
                     onTap: () async {
-                      await widget.onTap
-                          ?.let((onTap) => onTap(country, context));
+                      await widget.onTap?.let((onTap) => onTap(item, context));
 
                       if (!context.mounted) return;
                       Navigator.pop(context);