+++ packages/espressocash_app/test/migrations_test.dart
@@ -3,6 +3,7 @@ import 'package:drift_dev/api/migrations.dart';
 import 'package:espressocash_app/data/db/db.dart';
 import 'package:flutter_test/flutter_test.dart';
 
+// Run `make flutter_generate_test_schemas` to generate the file.
 import 'generated/schema.dart';
 
 void main() {