+++ packages/solana/lib/src/subscription_client/subscribed_message.g.dart
@@ -8,8 +8,8 @@ part of 'subscribed_message.dart';
 
 SubscribedMessage _$SubscribedMessageFromJson(Map<String, dynamic> json) =>
     SubscribedMessage(
-      result: json['result'] as int,
-      id: json['id'] as int,
+      result: (json['result'] as num).toInt(),
+      id: (json['id'] as num).toInt(),
     );
 
 Map<String, dynamic> _$SubscribedMessageToJson(SubscribedMessage instance) =>